Reverted renet implementation.

+++ src/space/core/atmospherics/events.rs
@@ -1,16 +1,16 @@
 use crate::space::core::networking::resources::{NetMessageType, ReliableServerMessage};
 
 pub struct NetMapDisplayAtmospherics {
-    pub handle: u64,
+    pub handle: u32,
     pub message: NetMessageType,
 }
 
 pub struct NetMapHoverAtmospherics {
-    pub handle: u64,
+    pub handle: u32,
     pub message: NetMessageType,
 }
 
 pub struct NetAtmosphericsNotices {
-    pub handle: u64,
+    pub handle: u32,
     pub message: ReliableServerMessage,
 }
