+++ src/space/core/connected_player/functions/player_selector_to_entities.rs
@@ -8,7 +8,7 @@ use crate::space::core::{
 
 pub fn player_selector_to_entities(
     command_executor_entity: Entity,
-    command_executor_handle_option: Option<u64>,
+    command_executor_handle_option: Option<u32>,
     player_selector_input: &str,
     used_names: &mut ResMut<UsedNames>,
     net_console_commands: &mut EventWriter<NetConsoleCommands>,
