+++ src/space/core/tab_actions/resources.rs
@@ -11,7 +11,7 @@ pub struct QueuedTabActions {
 
 pub struct QueuedTabAction {
     pub tab_id: String,
-    pub handle_option: Option<u64>,
+    pub handle_option: Option<u32>,
     pub target_cell_option: Option<(GridMapType, i16, i16, i16)>,
     pub target_entity_option: Option<u64>,
     pub belonging_entity_option: Option<u64>,
